requirements.txt
absl-py
astor
bleach
bottle
click
certifi
chardet
flask
future
gast
grpcio
h5py
html5lib
itsdangerous
idna
Jinja2
line-bot-sdk
Markdown
MarkupSafe
numpy
oauthlib
pillow
protobuf
requests
six
tensorboard
tensorboard-data-server
tensorflow
termcolor
urllib3
Werkzeug
wheel